Add test for POST retries
