+++ tests/test_chembl_client.py
@@ -11,7 +11,7 @@ from cachetools import TTLCache  # type: ignore[import-untyped]
 
 import library.rate_limiter as rl
 from library.chembl_client import ChemblClient
-from library.config import ApiCfg, RetryCfg, session_with_retry
+from library.config import ApiCfg, RetryCfg
 
 
 class DummyResponse:
@@ -206,19 +206,4 @@ def test_clients_do_not_share_cache() -> None:
     responses.add(responses.GET, url, json={"ok": 2}, status=200)
     client2 = ChemblClient(ApiCfg(), RetryCfg())
     client2.request_json(url, cfg=ApiCfg())
-    assert len(responses.calls) == 2
-
-
-@responses.activate
-def test_session_with_retry_retries_post() -> None:
-    """Ensure POST requests are retried when configured."""
-
-    url = "http://example.com/post"
-    responses.add(responses.POST, url, status=500)
-    responses.add(responses.POST, url, json={"ok": True}, status=200)
-
-    session = session_with_retry(ApiCfg(), RetryCfg(max_attempts=2, backoff_factor=0))
-    response = session.post(url)
-
-    assert response.status_code == 200
     assert len(responses.calls) == 2
